LPS-80570 Create a FormValuesUtil and small refactors on form resources
